+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es2017",
+    "target": "esnext",
     "sourceMap": true,
-    "moduleResolution": "node"
+    "moduleResolution": "node",
+    "allowJs": true
   },
   "exclude": [
     "node_modules"
-  ],
-  "extends": "tslint-config-standard"
+  ]
 }
